microservices/entities_service/requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.1.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.8
Django==4.2
django-cors-headers==4.3.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.13.1
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kombu==5.3.4
MarkupSafe==2.1.3
numpy==1.26.4
openapi-codec==1.3.2
packaging==23.2
pandas==2.2.0
Pillow==10.1.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.0
requests==2.31.0
rpds-py==0.15.1
setuptools==69.0.2
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.12
wheel==0.42.0
